// src/api/models/dailyStoreOperationResult.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Allta Partners
 * Allta Partners 서버 문서입니다.
 * OpenAPI spec version: 1.0
 */
import type { DailyStoreOperationCashHistoryItem } from './dailyStoreOperationCashHistoryItem';

export interface DailyStoreOperationResult {
  /** ID */
  id: string;
  /** 총 매출 */
  totalSales: number;
  /** 카드 매출 */
  cardSales: number;
  /** 현금 매출 */
  cashSales: number;
  /** 기타 매출 */
  otherSales: number;
  /**
   * 예상 현금 매출
   * @nullable
   */
  expectedCashSales?: number | null;
  /**
   * 시재금 차이
   * @nullable
   */
  cashDifference?: number | null;
  /**
   * 세차 횟수
   * @nullable
   */
  serviceCount?: number | null;
  /**
   * 예상 세차 횟수
   * @nullable
   */
  expectedServiceCount?: number | null;
  /**
   * 세차 횟수 차이
   * @nullable
   */
  serviceDifference?: number | null;
  /** 수정일 */
  openedAt: string;
  /**
   * 삭제일
   * @nullable
   */
  closedAt?: string | null;
  /** 현금 시재내역 목록 */
  cashHistories: DailyStoreOperationCashHistoryItem[];
}